examples/requirements-example.txt
matplotlib==3.4.1
datasets==2.6.1
torchvision==0.10.0
torchtext==0.14.0